requirements.txt
anyio==3.3.0
attrs==21.2.0
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
coverage==5.5
curlify==2.2.1
decorator==5.0.9
deepdiff==5.5.0
distlib==0.3.2
filelock==3.0.12
graphql-core==3.1.6
hypothesis==6.17.4
hypothesis-graphql==0.5.1
hypothesis-jsonschema==0.20.1
idna==3.2
iniconfig==1.1.1
jsonpath-rw==1.4.0
jsonpath-rw-ext==1.2.2
jsonschema==3.2.0
junit-xml==1.9
multidict==5.1.0
ordered-set==4.0.2
packaging==21.0
pluggy==1.0.0
pbr==5.6.0
platformdirs==2.3.0
pluggy==0.13.1
ply==3.11
py==1.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
pytest-subtests==0.5.0
PyYAML==5.4.1
requests==2.26.0
robotframework==4.1
robotframework-jsonlibrary==0.3.1
schemathesis==3.9.7
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
starlette==0.16.0
toml==0.10.2
tox==3.0.0
typing-extensions==3.10.0.2
urllib3==1.26.6
virtualenv==20.7.2
Werkzeug==2.0.1
yarl==1.6.3
